+++ elasticai/creator/examples/translate_linear_model.py
@@ -1,11 +1,11 @@
 import sys
 
 import torch
-
-from elasticai.creator.vhdl.number_representations import FixedPoint, FixedPointFactory
 from qat.hard_sigmoid import FixedPointHardSigmoid
-from elasticai.creator.vhdl.modules.linear import FixedPointLinear
 from qat.relu import FixedPointReLU
+
+from elasticai.creator.vhdl.modules.linear import FixedPointLinear
+from elasticai.creator.vhdl.number_representations import FixedPoint, FixedPointFactory
 from elasticai.creator.vhdl.translator.abstract.layers.fp_hard_sigmoid_module import (
     FPHardSigmoidTranslationArgs,
 )
@@ -67,7 +67,6 @@ def get_custom_build_mapping() -> BuildFunctionMapping:
 
 
 def main() -> None:
-
     torch.manual_seed(22)  # make result reproducible
 
     if len(sys.argv) < 2:
@@ -79,19 +78,8 @@ def main() -> None:
 
     model = FixedPointModel(fixed_point_factory)
 
-    translation_args = dict(
-        FixedPointLinear=FPLinear1dTranslationArgs(
-            fixed_point_factory=fixed_point_factory, work_library_name="work"
-        ),
-        FixedPointHardSigmoid=FPHardSigmoidTranslationArgs(
-            fixed_point_factory=fixed_point_factory
-        ),
-        FixedPointReLU=FPReLUTranslationArgs(fixed_point_factory=fixed_point_factory),
-    )
-
     code_repr = translator.translate_model(
         model=model,
-        translation_args=translation_args,
     )
 
     translator.save_code(code_repr=code_repr, path=build_path)