+++ elasticai/creator/tests/integration/vhdl/vhdl_file_test_case.py
@@ -2,7 +2,7 @@ import unittest
 from itertools import filterfalse
 
 from elasticai.creator import resource_utils
-from vhdl.code import Code
+from elasticai.creator.vhdl.code import Code
 
 
 class VHDLFileTestCase(unittest.TestCase):