+++ elasticai/creator/tests/unit/vhdl/test_translatable_root_layer_test.py
@@ -1,6 +1,6 @@
 import unittest
 
-from vhdl.hw_equivalent_layers.layers import AbstractTranslatableLayer
+from elasticai.creator.vhdl.hw_equivalent_layers.layers import AbstractTranslatableLayer
 
 
 class MyRootLayer(AbstractTranslatableLayer):