+++ elasticai/creator/tests/unit/vhdl/translator/abstract/layers/test_fp_hard_sigmoid_module.py
@@ -1,8 +1,6 @@
 import unittest
 
-from vhdl.code_files.fp_hard_sigmoid_file import (
-    FPHardSigmoidFile,
-)
+from elasticai.creator.vhdl.code_files.fp_hard_sigmoid_file import FPHardSigmoidFile
 from elasticai.creator.vhdl.number_representations import FixedPoint
 from elasticai.creator.vhdl.translator.abstract.layers.fp_hard_sigmoid_module import (
     FPHardSigmoidModule,