+++ elasticai/creator/tests/unit/vhdl/translator/test_build_function_mapping.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 from typing import Any
 
 import elasticai.creator.nn.layers as qtorch
-from vhdl.code import Code, CodeFile, CodeModule
+from elasticai.creator.vhdl.code import Code, CodeFile, CodeModule
 from elasticai.creator.vhdl.translator.build_function_mapping import (
     BuildFunctionMapping,
 )