+++ elasticai/creator/vhdl/code_files/fp_linear_1d_component.py
@@ -2,13 +2,13 @@ from dataclasses import dataclass, field
 from importlib.resources import read_text
 from typing import Callable
 
-from vhdl.code_files.utils import (
+from elasticai.creator.vhdl.code import Code, CodeFile
+from elasticai.creator.vhdl.code_files.utils import (
     calculate_address_width,
     derive_fixed_point_params_from_factory,
 )
-from vhdl.code import Code, CodeFile
 from elasticai.creator.vhdl.number_representations import FixedPoint
-from vhdl.vhdl_files import expand_template
+from elasticai.creator.vhdl.vhdl_files import expand_template
 
 
 @dataclass
@@ -17,8 +17,8 @@ class FPLinear1dFile(CodeFile):
         raise NotImplementedError()
 
     layer_id: str  # used to distinguish layers in the same model
-    in_features: int
-    out_features: int
+    in_feature_num: int
+    out_feature_num: int
     fixed_point_factory: Callable[[float], FixedPoint]
     work_library_name: str = field(default="work")
     resource_option: str = "auto"
@@ -27,14 +27,26 @@ class FPLinear1dFile(CodeFile):
         self.data_width, self.frac_width = derive_fixed_point_params_from_factory(
             self.fixed_point_factory
         )
-        self.x_addr_width = calculate_address_width(self.in_features)
-        self.y_addr_width = calculate_address_width(self.out_features)
+        self.x_addr_width = calculate_address_width(self.in_feature_num)
+        self.y_addr_width = calculate_address_width(self.out_feature_num)
+
+    def _template_parameters(self) -> dict[str, str]:
+        return dict(
+            (key, str(getattr(self, key)))
+            for key in (
+                "data_width",
+                "frac_width",
+                "x_addr_width",
+                "y_addr_width",
+                "in_feature_num",
+                "out_feature_num",
+            )
+        )
 
     @property
     def name(self) -> str:
         return f"fp_linear_1d_{self.layer_id}.vhd"
 
-    @property
     def code(self) -> Code:
         template = read_text("elasticai.creator.vhdl.templates", "fp_linear_1d.tpl.vhd")
 
@@ -42,13 +54,7 @@ class FPLinear1dFile(CodeFile):
             template.splitlines(),
             layer_name=self.layer_id,
             work_library_name=self.work_library_name,
-            data_width=self.data_width,
-            frac_width=self.frac_width,
-            x_addr_width=self.x_addr_width,
-            y_addr_width=self.y_addr_width,
-            in_feature_num=self.in_features,
-            out_feature_num=self.out_features,
             resource_option=f'"{self.resource_option}"',
+            **self._template_parameters(),
         )
-
-        yield from code
+        return code