+++ elasticai/creator/vhdl/number_representations.py
@@ -11,20 +11,17 @@ def _assert_range(value: float, total_bits: int, frac_bits: int) -> None:
     min_value = max_value * (-1)
     if not min_value <= value < max_value:
         raise ValueError(
-            (
-                f"Value {value} cannot represented as a fixed point value with {total_bits} total bits "
-                f"and {frac_bits} fraction bits (value range: [{min_value}, {max_value}))."
-            )
+            f"Value {value} cannot represented as a fixed point value with"
+            f" {total_bits} total bits and {frac_bits} fraction bits (value range:"
+            f" [{min_value}, {max_value}))."
         )
 
 
 def _assert_is_compatible(fp1: "FixedPoint", fp2: "FixedPoint") -> None:
     if not (fp1.total_bits == fp2.total_bits and fp1.frac_bits == fp2.frac_bits):
         raise ValueError(
-            (
-                f"FixedPoint objects not compatible (total_bits: {fp1.total_bits} != {fp2.total_bits}); "
-                f"frac_bits: {fp1.frac_bits} != {fp2.frac_bits})."
-            )
+            f"FixedPoint objects not compatible (total_bits: {fp1.total_bits} !="
+            f" {fp2.total_bits}); frac_bits: {fp1.frac_bits} != {fp2.frac_bits})."
         )
 
 
@@ -135,7 +132,10 @@ class FixedPoint:
         return str(int(self))
 
     def __repr__(self) -> str:
-        return f"FixedPoint(value={self._value}, total_bits={self._total_bits}, frac_bits={self._frac_bits})"
+        return (
+            f"FixedPoint(value={self._value}, total_bits={self._total_bits},"
+            f" frac_bits={self._frac_bits})"
+        )
 
     def _identical_fixed_point(self, value: float) -> "FixedPoint":
         return FixedPoint(
@@ -151,7 +151,8 @@ class FixedPoint:
     def from_unsigned_int(value: int, total_bits: int, frac_bits: int) -> "FixedPoint":
         if value > 2**total_bits - 1:
             raise ValueError(
-                f"Value {value} cannot interpreted as a fixed point with {total_bits} total bits."
+                f"Value {value} cannot interpreted as a fixed point with"
+                f" {total_bits} total bits."
             )
         is_negative = value & (1 << total_bits - 1) > 0
         if is_negative:
@@ -166,7 +167,7 @@ class FixedPoint:
         return FixedPoint(float_value, total_bits=total_bits, frac_bits=frac_bits)
 
     @staticmethod
-    def get_factory(total_bits: int, frac_bits: int) -> Callable[[float], "FixedPoint"]:
+    def get_factory(total_bits: int, frac_bits: int) -> "FixedPointFactory":
         return _FixedPointFactoryImpl(
             total_bits=total_bits, frac_bits=frac_bits, constructor=FixedPoint
         )
@@ -211,7 +212,10 @@ class ClippedFixedPoint(FixedPoint):
         )._value
 
     def __repr__(self) -> str:
-        return f"ClippedFixedPoint(value={self._value}, total_bits={self._total_bits}, frac_bits={self._frac_bits})"
+        return (
+            f"ClippedFixedPoint(value={self._value}, total_bits={self._total_bits},"
+            f" frac_bits={self._frac_bits})"
+        )
 
     def _identical_fixed_point(self, value: float) -> "ClippedFixedPoint":
         return ClippedFixedPoint(
@@ -260,7 +264,8 @@ def infer_total_and_frac_bits(*values: Sequence[FixedPoint]) -> tuple[int, int]:
     for value in chain(*values):
         if value.total_bits != total_bits or value.frac_bits != frac_bits:
             raise ValueError(
-                "Cannot infer total bits and frac bits from a list with mixed total bits or frac bits."
+                "Cannot infer total bits and frac bits from a list with mixed total"
+                " bits or frac bits."
             )
     return total_bits, frac_bits